+++ R/eucast.R
@@ -20,7 +20,7 @@
 #'
 #' Apply expert rules (like intrinsic resistance), as defined by the European Committee on Antimicrobial Susceptibility Testing (EUCAST, \url{http://eucast.org}), see \emph{Source}.
 #' @param tbl table with antibiotic columns, like e.g. \code{amox} and \code{amcl}
-#' @param col_mo column name of the bacteria ID in \code{tbl} - values of this column should be present in \code{microorganisms$mo}, see \code{\link{microorganisms}}
+#' @param col_mo column name of the microbial ID in \code{tbl} - values in this column should be present in \code{microorganisms$mo}, see \code{\link{microorganisms}}
 #' @param info print progress
 #' @param amcl,amik,amox,ampi,azit,azlo,aztr,cefa,cfep,cfot,cfox,cfra,cfta,cftr,cfur,chlo,cipr,clar,clin,clox,coli,czol,dapt,doxy,erta,eryt,fosf,fusi,gent,imip,kana,levo,linc,line,mero,mezl,mino,moxi,nali,neom,neti,nitr,norf,novo,oflo,peni,pita,poly,pris,qida,rifa,roxi,siso,teic,tetr,tica,tige,tobr,trim,trsu,vanc column name of an antibiotic. Use \code{NA} to skip a column, like \code{tica = NA}. Non-existing columns will anyway be skipped. See the Antibiotics section for an explanation of the abbreviations.
 #' @param col_bactid Deprecated. Use \code{col_mo} instead.
@@ -103,11 +103,12 @@
 #'   \url{http://www.eucast.org/fileadmin/src/media/PDFs/EUCAST_files/Expert_Rules/Expert_rules_intrinsic_exceptional_V3.1.pdf}
 #' @examples
 #' a <- EUCAST_rules(septic_patients)
-#' a <- data.frame(mo = c("STAAUR",  # Staphylococcus aureus
-#'                        "ENCFAE",  # Enterococcus faecalis
-#'                        "ESCCOL",  # Escherichia coli
-#'                        "KLEPNE",  # Klebsiella pneumoniae
-#'                        "PSEAER"), # Pseudomonas aeruginosa
+#'
+#' a <- data.frame(mo = c("Staphylococcus aureus",
+#'                        "Enterococcus faecalis",
+#'                        "Escherichia coli",
+#'                        "Klebsiella pneumoniae",
+#'                        "Pseudomonas aeruginosa"),
 #'                 vanc = "-",       # Vancomycin
 #'                 amox = "-",       # Amoxicillin
 #'                 coli = "-",       # Colistin
@@ -116,7 +117,7 @@
 #'                 stringsAsFactors = FALSE)
 #' a
 #'
-#' b <- EUCAST_rules(a)
+#' b <- EUCAST_rules(a) # 11 results will be forced as R due to intrinsic resistance
 #' b
 EUCAST_rules <- function(tbl,
                          col_mo = 'mo',